% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTables.R
\name{compareTables}
\alias{compareTables}
\title{Compares the structure of relational schema from two databases held in
matrices}
\usage{
compareTables(older_db_mat, newer_db_mat)
}
\arguments{
\item{older_db_mat}{A matrix containing table information from a newer database}

\item{newer_db_mat}{A matrix containing table information from an older database}
}
\value{
A vector containing strings representing information about tables in markdown format
}
\description{
Compares the structure of relational schema from two databases held in
matrices
}
