% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewTables.R
\name{getNewTables}
\alias{getNewTables}
\title{Returns the names of new tables between two lists of table names from two versions of a database}
\usage{
getNewTables(older_db_mat_tab_names, newer_db_mat_tab_names)
}
\arguments{
\item{older_db_mat_tab_names}{A vector containing table names from an  older database}

\item{newer_db_mat_tab_names}{A vector containing table names from a newer database}
}
\value{
A vector containing table names unique to the newer database
}
\description{
Returns the names of new tables between two lists of table names from two versions of a database
}
