% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableInfo.R
\name{getTableInfo}
\alias{getTableInfo}
\title{Extracts the column names for tables in a PostgreSQL SQL dump file.}
\usage{
getTableInfo(sqlfilepath)
}
\arguments{
\item{sqlfilepath}{A file path to the SQL dump file.}
}
\value{
A matrix containing the names of database tables and their columns.
}
\description{
Extracts the column names for tables in a PostgreSQL SQL dump file.
}
\examples{
\dontrun{
getTableInfo("sikfaan/inst/testdata/older.sql")
}
}
