% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableNames.R
\name{getTableNames}
\alias{getTableNames}
\title{Extract table names from PostgreSQL SQL dump file.}
\usage{
getTableNames(sqlfilepath)
}
\arguments{
\item{sqlfilepath}{A file path to the SQL dump file.}
}
\value{
A vector containing a list of table names.
}
\description{
Extract table names from PostgreSQL SQL dump file.
}
\examples{
\dontrun{
getTableNames("sikfaan/inst/testdata/older.sql")
}
}
