% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputTableMarkdown.R
\name{outputTableMarkdown}
\alias{outputTableMarkdown}
\title{Converts information about relation tables from two databases in matrices
into markdown format}
\usage{
outputTableMarkdown(table_name_list, newer_db_mat, older_db_mat, new_only)
}
\arguments{
\item{table_name_list}{A list of table names}

\item{newer_db_mat}{A matrix containing information about a newer database}

\item{older_db_mat}{A matrix containing information about an older database}

\item{new_only}{A boolean value to specify whether table information from databases are required or not}
}
\value{
A vector containing strings representing information about tables in markdown format
}
\description{
Converts information about relation tables from two databases in matrices
into markdown format
}
